mysite/requirements.txt
APScheduler==3.6.3
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
confusable-homoglyphs==3.2.0
Django==3.0.6
django-extensions==2.2.9
django-multi-email-field==0.6.1
django-registration==3.1
idna==2.9
psycopg2==2.8.5
pydotplus==2.0.2
pyparsing==2.4.7
python-decouple==3.3
pytz==2020.1
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
tzlocal==2.1
urllib3==1.25.9
